dividing main into separate files
